✨ Bloquage Compte

+++ src/main/java/fr/univlorraine/miage/revolutmiage/compte/infra/rest/CompteResource.java
@@ -54,7 +54,7 @@ public class CompteResource extends DefaultResource {
     public EntityModel<CompteDTO> getCompteById(@PathVariable final String iban) {
         final Optional<Compte> optionalCompte = catalog.findByIban(iban);
 
-        if (optionalCompte.isEmpty()) {
+        if (optionalCompte.isEmpty() || !optionalCompte.get().getUtilisateur().getNumeroPasseport().equals(currentUsername())) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
 
@@ -69,24 +69,44 @@ public class CompteResource extends DefaultResource {
 
     @GetMapping("{iban}/solde")
     public ResponseEntity<?> getCompteSolde(@PathVariable final String iban) {
-        return ResponseEntity.of(catalog.findByIban(iban).map(compte -> SimpleResponse.sendDouble("solde", compte.getSolde())));
+        final Optional<Compte> optionalCompte = catalog.findByIban(iban);
+
+        if (optionalCompte.isEmpty() || !optionalCompte.get().getUtilisateur().getNumeroPasseport().equals(currentUsername())) {
+            return ResponseEntity.notFound().build();
+        }
+
+        return ResponseEntity.of(optionalCompte.map(compte -> SimpleResponse.sendDouble("solde", compte.getSolde())));
     }
 
     @GetMapping("{iban}/cartes")
     public ResponseEntity<?> getCompteCartes(@PathVariable final String iban) {
-        return ResponseEntity.of(catalog.findByIban(iban).map(Compte::getCartes).map(carteMapper::toDtos));
+        final Optional<Compte> optionalCompte = catalog.findByIban(iban);
+
+        if (optionalCompte.isEmpty() || !optionalCompte.get().getUtilisateur().getNumeroPasseport().equals(currentUsername())) {
+            return ResponseEntity.notFound().build();
+        }
+
+        return ResponseEntity.of(optionalCompte.map(Compte::getCartes).map(carteMapper::toDtos));
     }
 
     @GetMapping("{iban}/cartes/{numeroCarte}")
     public ResponseEntity<?> getCompteCarte(@PathVariable final String iban, @PathVariable final String numeroCarte) {
-        return ResponseEntity.of(catalog.findByIban(iban)
+        final Optional<Compte> optionalCompte = catalog.findByIban(iban);
+
+        if (optionalCompte.isEmpty() || !optionalCompte.get().getUtilisateur().getNumeroPasseport().equals(currentUsername())) {
+            return ResponseEntity.notFound().build();
+        }
+
+        return ResponseEntity.of(optionalCompte
                 .map(Compte::getCartes).get().stream()
                 .filter(carte -> numeroCarte.equals(carte.getNumeroCarte())).findFirst().map(carteMapper::toDto));
     }
 
     @GetMapping("{iban}/operations")
     public ResponseEntity<List<OperationDTO>> getCompteOperations(@PathVariable final String iban, @RequestParam final Optional<String> categorie, @RequestParam final Optional<String> pays) {
-        if (catalog.findByIban(iban).isEmpty()) {
+        final Optional<Compte> optionalCompte = catalog.findByIban(iban);
+
+        if (optionalCompte.isEmpty() || !optionalCompte.get().getUtilisateur().getNumeroPasseport().equals(currentUsername())) {
             return ResponseEntity.notFound().build();
         }
 
@@ -109,19 +129,33 @@ public class CompteResource extends DefaultResource {
 
     @PostMapping("{iban}/cartes")
     public ResponseEntity<?> creerCompteCarte(@PathVariable final String iban, @RequestBody final UpdateCarteInput input) {
+        final Optional<Compte> optionalCompte = catalog.findByIban(iban);
+
+        if (optionalCompte.isEmpty() || !optionalCompte.get().getUtilisateur().getNumeroPasseport().equals(currentUsername())) {
+            return ResponseEntity.notFound().build();
+        }
+
         input.setCompteIban(iban);
         return carteResource.creerCarte(input);
     }
 
     @PutMapping("{iban}/cartes/{numeroCarte}")
     public ResponseEntity<?> modifierCompteCarte(@PathVariable final String iban, @PathVariable final String numeroCarte, @RequestBody final UpdateCarteInput input) {
+        final Optional<Compte> optionalCompte = catalog.findByIban(iban);
+
+        if (optionalCompte.isEmpty() || !optionalCompte.get().getUtilisateur().getNumeroPasseport().equals(currentUsername())) {
+            return ResponseEntity.notFound().build();
+        }
+
         input.setCompteIban(iban);
         return carteResource.modifierCarte(numeroCarte, input);
     }
 
     @DeleteMapping("{iban}")
     public ResponseEntity<?> supprimerCompte(@PathVariable final String iban) {
-        if (catalog.findByIban(iban).isEmpty()) {
+        final Optional<Compte> optionalCompte = catalog.findByIban(iban);
+
+        if (optionalCompte.isEmpty() || !optionalCompte.get().getUtilisateur().getNumeroPasseport().equals(currentUsername())) {
             return ResponseEntity.notFound().build();
         }
         deleteCompte.accept(new DeleteCompteInput().setIban(iban));
@@ -130,6 +164,12 @@ public class CompteResource extends DefaultResource {
 
     @DeleteMapping("{iban}/cartes/{numeroCarte}")
     public ResponseEntity<?> supprimerCompteCarte(@PathVariable final String iban, @PathVariable final String numeroCarte) {
+        final Optional<Compte> optionalCompte = catalog.findByIban(iban);
+
+        if (optionalCompte.isEmpty() || !optionalCompte.get().getUtilisateur().getNumeroPasseport().equals(currentUsername())) {
+            return ResponseEntity.notFound().build();
+        }
+
         return carteResource.supprimerCarte(numeroCarte);
     }
 }
